move fast_merge to raphtory-itertools
