--- fast_merge/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "fast_merge"
-version.workspace = true
-documentation.workspace = true
-repository.workspace = true
-license.workspace = true
-readme.workspace = true
-homepage.workspace = true
-keywords.workspace = true
-authors.workspace = true
-rust-version.workspace = true
-edition.workspace = true
-
-[dependencies]
-itertools = { workspace = true }
-
-[dev-dependencies]
-proptest.workspace = true
-criterion = { workspace = true }
-rand = { workspace = true }
-
-
-[[bench]]
-name = "bench"
-harness = false
