+++ raphtory-itertools/Cargo.toml
@@ -12,3 +12,14 @@ rust-version.workspace = true
 edition.workspace = true
 
 [dependencies]
+itertools = { workspace = true }
+
+[dev-dependencies]
+proptest.workspace = true
+criterion = { workspace = true }
+rand = { workspace = true }
+
+
+[[bench]]
+name = "bench"
+harness = false
