+++ raphtory-itertools/src/lib.rs
@@ -1,3 +1,6 @@
+pub(crate) mod merge;
+pub(crate) mod merge_impl;
 mod take;
 
+pub use merge::{FastMerge, FastMergeExt};
 pub use take::{ReTake, TakeExt};
